Refactor: Migrate to SootUp, 1-to-1 Evaluators, and Architecture Cleanup
- Migrated analysis engine to SootUp 2.0.0
- Refactored all evaluators to 1-to-1 mapping with graph-based traversal
- Reorganized evaluators into domain packages (jdbc, remote, async, architecture, concurrency)
- Renamed adapter.out.rule to adapter.out.persistence
- Implemented JAR file analysis support (auto-extraction)
- Removed JUnit and legacy analysis code

+++ src/main/java/com/vidnyan/ate/application/port/in/AnalyzeCodeUseCase.java
@@ -58,9 +58,9 @@ public interface AnalyzeCodeUseCase {
      * Analysis statistics.
      */
     record AnalysisStats(
-        int filesAnalyzed,
-        int typesAnalyzed,
-        int methodsAnalyzed,
+        long filesAnalyzed,
+        long typesAnalyzed,
+        long methodsAnalyzed,
         int rulesEvaluated,
         long totalDurationMs
     ) {}
